backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.3.5
openai-whisper==20231214
torch==2.1.2
librosa==0.10.0
numpy==1.24.3
scipy==1.11.4
requests==2.31.0
pydantic==2.5.0
speechbrain==0.5.15
resemblyzer==0.1.1.dev0
soundfile==0.12.1
twilio==8.10.0

